// apps/marketing/marketing-mfe/src/app/http/cafe/raw-customerproduct-holding/models/cafe-raw-customerproductholding-rawomapi-services-specification.model.ts
import { CafeRawCustomerproductholdingRawomapiLocalizedContentModel } from './cafe-raw-customerproductholding-rawomapi-localizedContent.model';

export class CafeRawCustomerproductholdingRawomapiServicesSpecificationModel {
  labelkey: string;
  value: string;
  maxvalue: string;
  minvalue: string;
  unit: string;
  visible: boolean;
  weight: string;
  localizedcontent: CafeRawCustomerproductholdingRawomapiLocalizedContentModel[];
}