// apps/marketing/marketing-mfe/src/app/http/cafe/raw-customerproduct-holding/models/cafe-raw-customerproductholding-rawomapi-characteristics.model.ts
import { CafeRawCustomerproductholdingRawomapiCharactersiticsUnitvalue } from './cafe-raw-customerproductholding-rawomapi-charactersitics-unitvalue';
import { RawConstituentInterface } from '@telenet/ng-lib-omapi';

export class CafeRawCustomerproductholdingRawomapiCharacteristicsModel {
  alert_threshold_marker: CafeRawCustomerproductholdingRawomapiCharactersiticsUnitvalue;
  initial_threshold_1: CafeRawCustomerproductholdingRawomapiCharactersiticsUnitvalue;
  initial_threshold_2: CafeRawCustomerproductholdingRawomapiCharactersiticsUnitvalue;
  alert_threshold: CafeRawCustomerproductholdingRawomapiCharactersiticsUnitvalue;
  service_category_limit: CafeRawCustomerproductholdingRawomapiCharactersiticsUnitvalue;
  sharedmobiledata: CafeRawCustomerproductholdingRawomapiCharactersiticsUnitvalue;
  productsegment: string;
  producttier: string;
  productgroup: string;
  productsubgroup: string;
  max_mobile_lines: string;
  activationsalesofferingid: string;
  max_mobile_includeddata_reminders: string;
  max_mobile_outofbundle_reminders: string;
  productlevel: string;
  salespricevatexcl: CafeRawCustomerproductholdingRawomapiCharactersiticsUnitvalue;
  salespricevatincl: CafeRawCustomerproductholdingRawomapiCharactersiticsUnitvalue;
  rentalpricevatexcl: CafeRawCustomerproductholdingRawomapiCharactersiticsUnitvalue;
  rentalpricevatincl: CafeRawCustomerproductholdingRawomapiCharactersiticsUnitvalue;
  service_category: string;
  detailed_scale: CafeRawCustomerproductholdingRawomapiCharactersiticsUnitvalue;
  sharedmobiledata_unlimited: CafeRawCustomerproductholdingRawomapiCharactersiticsUnitvalue;
  huwrecipientsmandatory: string;
  devicetype: string;
  is_sellable: string;
  weightnl: string;
  weightfr: string;
  constituents: RawConstituentInterface[];
}