// apps/marketing/marketing-mfe/src/app/http/cafe/raw-customerproduct-holding/models/cafe-raw-customerproductholding-rawomapi-subcategory.model.ts
import { CafeRawCustomerproductholdingRawomapiLocalizedContentModel } from './cafe-raw-customerproductholding-rawomapi-localizedContent.model';

export class CafeRawCustomerproductholdingRawomapiSubcategoryModel {
  name: string;
  localizedcontent: CafeRawCustomerproductholdingRawomapiLocalizedContentModel[];
  weight: number;
  mutualexclusion: boolean;
}